Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> Thanks Roni.  We will integrated this to RIC Platform. 
> How are the ports and hostnames provided to the tracing clients and which RIC components run the tracing clients?

The idea was that there would be a config map which defines environment variables that contains  the address and port and other tracing configuration values. Client can then include that config map in their kubernetes configuration. There is an example yaml, env.yaml in the templates directory which does this.

I'm not sure which components will the first to use tracing. I assume E2 termination and maybe some xapps?

Patch-set: 1
Label: Code-Review=0